+++ CHANGELOG.md
@@ -13,6 +13,16 @@ As this project is a user-facing application, the places in the semantic version
 
 ## [Unreleased](https://github.com/spraakbanken/mink-frontend/compare/v1.16.1...HEAD)
 
+### Added
+
+- View result file [#170](https://github.com/spraakbanken/mink-frontend/issues/170)
+- Explain export file types [#192](https://github.com/spraakbanken/mink-frontend/issues/192)
+- List allowed source file extensions on upload area [#234](https://github.com/spraakbanken/mink-frontend/issues/234)
+
+### Changed
+
+- Removed source plaintext view [#227](https://github.com/spraakbanken/mink-frontend/issues/227)
+
 ## [1.16.1](https://github.com/spraakbanken/mink-frontend/releases/tag/v1.16.1) (2026-02-16)
 
 ### Fixed